+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-m5kdev
 
+## 0.16.4
+
+### Patch Changes
+
+- colors sidebar
+
+## 0.16.3
+
+### Patch Changes
+
+- dropdown
+
 ## 0.16.2
 
 ### Patch Changes