+++ packages/commons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @m5kdev/commons
 
+## 0.16.4
+
+### Patch Changes
+
+- colors sidebar
+
+## 0.16.3
+
+### Patch Changes
+
+- dropdown
+
 ## 0.16.2
 
 ### Patch Changes