+++ packages/frontend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @m5kdev/frontend
 
+## 0.16.4
+
+### Patch Changes
+
+- colors sidebar
+- Updated dependencies
+  - @m5kdev/commons@0.16.4
+
+## 0.16.3
+
+### Patch Changes
+
+- dropdown
+- Updated dependencies
+  - @m5kdev/commons@0.16.3
+
 ## 0.16.2
 
 ### Patch Changes